\name{gitSubmoduleAdd}
\alias{gitSubmoduleAdd}
\title{Git add submodule}
\usage{
  gitSubmoduleAdd(url, dir, path = c())
}
\arguments{
  \item{url}{submodule url}

  \item{dir}{repository directory}

  \item{path}{submodule path (optional)}
}
\value{
  git output
}
\description{
  Add a submodule.
}

