\name{gitSubmoduleStatus}
\alias{gitSubmoduleStatus}
\title{Get Submodule Status String}
\usage{
  gitSubmoduleStatus(directory, submodules)
}
\arguments{
  \item{directory}{repository directory}

  \item{submodules}{requested submodules}
}
\description{
  Get informative submodule status string from git status
}

