\name{showMessage}
\alias{showMessage}
\title{Show Dialog}
\usage{
  showMessage(..., type = "info", obj = obj)
}
\arguments{
  \item{...}{marked up message to display}

  \item{type}{of message, e.g., "info", "error", ...}

  \item{obj}{gitR object}
}
\description{
  Shows a dbraialog while blocking the input to the parent
  window.
}

